+++ src/gt4py/next/program_processors/runners/dace/gtir_builtin_translators.py
@@ -30,7 +30,7 @@ from gt4py.next.program_processors.runners.dace import (
     gtir_sdfg_utils,
     utils as gtx_dace_utils,
 )
-from gt4py.next.program_processors.runners.dace.gtir_scan_translator import translate_scan
+from gt4py.next.program_processors.runners.dace.gtir_lower_scan import translate_scan
 from gt4py.next.type_system import type_info as ti, type_specifications as ts
